Add files via upload
A software system is being developed that can predict multiple diseases using machine learning 
algorithms. The system will be able to predict heart disease, Diabetes. The system will be designed 
as a web application and will be implemented using Python. 
The system will be used by both individuals and healthcare professionals to predict and manage 
the risk of disease. Individuals will be able to use the system to assess their own risk of disease, 
and healthcare professionals will be able to use the system to predict the risk of disease for their 
patients. 
The system is expected to have a significant impact on healthcare by helping to detect diseases 
early and prevent serious complications. The system will also help to reduce the cost of healthcare 
by reducing the need for unnecessary tests and procedures.